requirements.txt
SpeechRecognition
pyttsx3
PyAudio
streamlit
pipwin
gTTs
google-generativeai